Add files via upload
Step 1
Remove the Reset button and its functionality.

Step 2
Add the functionality which counts the number of clicks made to both the increment and decrement buttons. - This number should show on the page with a label.

Step 3
Add an input field where the user can input a number. Next to the input should be a button. When the button is clicked, the original increment and decrement count should reset to the number that was inputted and the number of clicks should reset to zero. - The number should not add onto the number that was created by clicking the increment and decrement buttons. The number listed should show the exact number inputted.